+++ window/src/os/macos/mod.rs
@@ -7,6 +7,7 @@ use objc::*;
 mod app;
 pub mod bitmap;
 pub mod connection;
+pub mod menu;
 pub mod window;
 
 mod keycodes;