+++ window/src/os/macos/window.rs
@@ -4,6 +4,7 @@
 use super::keycodes::*;
 use super::{nsstring, nsstring_to_str};
 use crate::connection::ConnectionOps;
+use crate::os::macos::menu::{MenuItem, RepresentedItem};
 use crate::parameters::{Border, Parameters, TitleBar};
 use crate::{
     Clipboard, Connection, DeadKeyStatus, Dimensions, Handled, KeyCode, KeyEvent, Modifiers,
@@ -1903,6 +1904,28 @@ impl WindowView {
         NO
     }
 
+    extern "C" fn wezterm_perform_key_assignment(
+        this: &mut Object,
+        _sel: Sel,
+        menu_item: *mut Object,
+    ) {
+        let menu_item = MenuItem::with_menu_item(menu_item);
+        // Safe because weztermPerformKeyAssignment: is only used with KeyAssignment
+        let action = menu_item.get_represented_item();
+        log::debug!("wezterm_perform_key_assignment {action:?}",);
+        match action {
+            Some(RepresentedItem::KeyAssignment(action)) => {
+                if let Some(this) = Self::get_this(this) {
+                    this.inner
+                        .borrow_mut()
+                        .events
+                        .dispatch(WindowEvent::PerformKeyAssignment(action));
+                }
+            }
+            None => {}
+        }
+    }
+
     extern "C" fn window_will_close(this: &mut Object, _sel: Sel, _id: id) {
         if let Some(this) = Self::get_this(this) {
             // Advise the window of its impending death
@@ -2708,6 +2731,12 @@ impl WindowView {
             cls.add_method(
                 sel!(dealloc),
                 WindowView::dealloc as extern "C" fn(&mut Object, Sel),
+            );
+
+            cls.add_method(
+                sel!(weztermPerformKeyAssignment:),
+                Self::wezterm_perform_key_assignment
+                    as extern "C" fn(&mut Object, Sel, *mut Object),
             );
 
             cls.add_method(